requirements.txt
pandas==1.2.3
python-dotenv==0.15.0
